requirements.txt
python-telegram-bot>=21.4
pydantic-settings==2.0.3
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
yt-dlp
